Improve string initialization sequence
This change open-codes `string-set!` in compiler-generated string
initialization code. In addition, it changes the `%store-mem` code
generation to be able to insert constant offsets into the `i32.store`
offset field, rather than having to do addition on the arguments at
runtime.

Compile time performance is basically unchanged, but this leads to a
7% decrease in compiler binary size.